+++ lib/mixins/path_finding.dart
@@ -66,7 +66,6 @@ mixin PathFinding on Movement {
     Vector2 position, {
     List? ignoreCollisions,
     VoidCallback? onFinish,
-    bool pathLineEnabled = true,
   }) async {
     if (!hasGameRef) {
       return Future.value([]);
@@ -82,9 +81,7 @@ mixin PathFinding on Movement {
         ignoreCollisions: ignoreCollisions,
       ),
     );
-    if (pathLineEnabled) {
-      _addLinePathComponent();
-    }
+    _addLinePathComponent();
 
     return _currentPath;
   }
@@ -92,7 +89,6 @@ mixin PathFinding on Movement {
   void moveAlongThePath(
     List<Vector2> path, {
     VoidCallback? onFinish,
-    bool pathLineEnabled = true,
   }) {
     if (!hasGameRef) {
       return;
@@ -103,9 +99,7 @@ mixin PathFinding on Movement {
     _removeLinePathComponent();
 
     _currentPath = path;
-    if (pathLineEnabled) {
-      _addLinePathComponent();
-    }
+    _addLinePathComponent();
   }
 
   List<Vector2> getPathToPosition(