+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT
 - Adds `moveAlongThePath` method in `PathFinding` mixin.
+- Bugfix in `Vision` mixin.
 
 # 3.6.0
 - Adds param `hudComponents` in `BonfireWidget`;